Consolidate profile-aware logger tests
- Keep logger override and fallback coverage with the profile-aware
  logger tests
- Remove the duplicated legacy core logger test file
